requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==24.3.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
distro==1.9.0
fastapi==0.115.6
Flask==3.0.3
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-speech==2.29.0
google-cloud-texttospeech==2.23.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
openai==0.28.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
python-dotenv==1.0.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
twilio==9.4.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn[standard]==0.34.0
Werkzeug==3.0.6
yarl==1.18.3
websockets==11.0.3
